refactor: 엔티티 domain 폴더 밑으로 위치 수정
