+++ src/test/java/TeamDustKGU/dustbackend/board/domain/BoardTest.java
@@ -1,7 +1,7 @@
 package TeamDustKGU.dustbackend.board.domain;
 
-import TeamDustKGU.dustbackend.board.Board;
-import TeamDustKGU.dustbackend.comment.Comment;
+import TeamDustKGU.dustbackend.board.domain.Board;
+import TeamDustKGU.dustbackend.comment.domain.Comment;
 import TeamDustKGU.dustbackend.user.domain.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
