+++ src/test/java/TeamDustKGU/dustbackend/fixture/BoardFixture.java
@@ -1,6 +1,6 @@
 package TeamDustKGU.dustbackend.fixture;
 
-import TeamDustKGU.dustbackend.board.Board;
+import TeamDustKGU.dustbackend.board.domain.Board;
 import TeamDustKGU.dustbackend.user.domain.User;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
