+++ src/test/java/TeamDustKGU/dustbackend/fixture/CommentFixture.java
@@ -1,7 +1,7 @@
 package TeamDustKGU.dustbackend.fixture;
 
-import TeamDustKGU.dustbackend.board.Board;
-import TeamDustKGU.dustbackend.comment.Comment;
+import TeamDustKGU.dustbackend.board.domain.Board;
+import TeamDustKGU.dustbackend.comment.domain.Comment;
 import TeamDustKGU.dustbackend.user.domain.User;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
